Add timer ingame.

+++ Assets/MyStuff/Scripts/Data/Data.cs
@@ -2,5 +2,5 @@ public static class Data
 {
     public static int currency             = 0 ;
     public static int maxSeconds           = 0 ;
-    public static int currentTimer         = 0 ;
+    public static float currentTimer         = 0 ;
 }
